Move helpers from utils to test_utils in tests
